Creation du Legacy service (amené a disparaitre) et de la branche developp
